refactor(application): 패키지명 오타 수정
작업 내용
- constant 패키지 이름 오타 수정
