+++ src/main/java/lotto/constant/OutputMessage.java
@@ -1,4 +1,4 @@
-package lotto.costant;
+package lotto.constant;
 
 public enum OutputMessage {
     PURCHASE_COMPLETE("%d개를 구매했습니다."),
