+++ src/main/java/lotto/constant/Prompt.java
@@ -1,4 +1,4 @@
-package lotto.costant;
+package lotto.constant;
 
 public enum Prompt {
     PURCHASE("구입금액을 입력해 주세요."),
