+++ src/main/java/lotto/controller/LottoController.java
@@ -1,14 +1,13 @@
 package lotto.controller;
 
+import java.util.List;
 import lotto.config.ApplicationConfiguration;
-import lotto.costant.WinningTier;
+import lotto.constant.WinningTier;
 import lotto.domain.Lotto;
 import lotto.domain.WinningLotto;
 import lotto.service.InputService;
 import lotto.service.LottoService;
 import lotto.service.OutputService;
-
-import java.util.List;
 
 public class LottoController {
