+++ src/main/java/lotto/service/OutputService.java
@@ -3,8 +3,8 @@ package lotto.service;
 import java.text.DecimalFormat;
 import java.util.Arrays;
 import java.util.List;
-import lotto.costant.OutputMessage;
-import lotto.costant.WinningTier;
+import lotto.constant.OutputMessage;
+import lotto.constant.WinningTier;
 import lotto.domain.Lotto;
 import lotto.domain.WinningCondition;
 import lotto.view.OutputView;
